+++ app/Http/Controllers/Admin/WhyChooseController.php
@@ -44,6 +44,11 @@ class WhyChooseController extends Controller
         $final_name = 'why-choose-'.$ai_id.'.'.$ext;
         $request->file('photo')->move(public_path('uploads/'), $final_name);
         $data['photo'] = $final_name;
+        if($request->has('industry_id')){
+            $data['industry_id']= json_encode($data['industry_id']);
+        }else{
+            $data['industry_id'] = null;
+        }
 
         $why_choose->fill($data)->save();
         return redirect()->route('admin.why_choose.index')->with('success', 'Why Choose Us Item is added successfully!');
@@ -52,8 +57,15 @@ class WhyChooseController extends Controller
     public function edit($id)
     {
         $why_choose = WhyChooseItem::findOrFail($id);
+        $industry_id = "";
+        if($why_choose->industry_id != null){
+            $industry_id = json_decode($why_choose->industry_id, true); 
+        }else{
+            $industry_id = [];
+        }
         $industry = Industry::all();
-        return view('admin.why_choose.edit', compact('why_choose','industry'));
+
+        return view('admin.why_choose.edit', compact('why_choose','industry','industry_id'));
     }
 
     public function update(Request $request, $id)
@@ -73,7 +85,9 @@ class WhyChooseController extends Controller
                 ],
                 'photo' => 'image|mimes:jpeg,png,jpg,gif|max:2048'
             ]);
-            unlink(public_path('uploads/'.$why_choose->photo));
+            if($request->input('current_photo') && file_exists($why_choose->photo)){
+                unlink(public_path('uploads/'.$why_choose->photo));
+            }
             $ext = $request->file('photo')->extension();
             $final_name = 'why-choose-'.$id.'.'.$ext;
             $request->file('photo')->move(public_path('uploads/'), $final_name);
@@ -88,6 +102,11 @@ class WhyChooseController extends Controller
             $data['photo'] = $why_choose->photo;
         }
 
+        if($request->has('industry_id')){
+            $data['industry_id']= json_encode($data['industry_id']);
+        }else{
+            $data['industry_id'] = null;
+        }
         $why_choose->fill($data)->save();
         return redirect()->route('admin.why_choose.index')->with('success', 'Why Choose Us Item is updated successfully!');
     }
